Fix integer overflow in ResourceMgr type list size calculations
Three instances in ResourceMgr calculated typeListSize without properly casting
numTypes to UInt32 before multiplication:

UInt32 typeListSize = 2 + (numTypes * sizeof(TypeListEntry));

Since numTypes is UInt16, this could theoretically overflow in the multiplication
before being assigned to UInt32, though in practice sizeof(TypeListEntry) = 8
makes actual overflow unlikely. However, similar code at lines 886 and 903 was
already fixed with proper casting.

Changed all three instances to match the fixed pattern:

UInt32 typeListSize = 2 + ((UInt32)numTypes * sizeof(TypeListEntry));

Locations: src/ResourceMgr/ResourceMgr.c:342, 548, 1307

+++ src/ResourceMgr/ResourceMgr.c
@@ -339,7 +339,7 @@ parse_resources:
                 UInt16 numTypes = read_be16(typeList) + 1;
 
                 /* Bounds check: ensure type list fits in map */
-                UInt32 typeListSize = 2 + (numTypes * sizeof(TypeListEntry));
+                UInt32 typeListSize = 2 + ((UInt32)numTypes * sizeof(TypeListEntry));
                 if (typeListOff + typeListSize > mapSize) {
                     serial_puts("[ResourceMgr] Warning: Type list exceeds map bounds\n");
                     gResMgr.resError = mapReadErr;
@@ -545,7 +545,7 @@ TypeListEntry* ResMap_FindType(ResFile* file, ResType type) {
     numTypes = read_be16(typeList) + 1;  /* Count is stored as n-1 */
 
     /* Bounds check: ensure type list entries fit in map */
-    UInt32 typeListSize = 2 + (numTypes * sizeof(TypeListEntry));
+    UInt32 typeListSize = 2 + ((UInt32)numTypes * sizeof(TypeListEntry));
     if (typeListOff + typeListSize > file->mapSize) {
         gResMgr.resError = mapReadErr;
         return NULL;
@@ -1304,7 +1304,7 @@ static ResType ResMap_GetIndType(ResFile* file, SInt16 index) {
     }
 
     /* Bounds check: ensure type list entries fit in map */
-    UInt32 typeListSize = 2 + (numTypes * sizeof(TypeListEntry));
+    UInt32 typeListSize = 2 + ((UInt32)numTypes * sizeof(TypeListEntry));
     if (typeListOff + typeListSize > file->mapSize) {
         gResMgr.resError = mapReadErr;
         return 0;
